Add 4D Barnes-Hut embedding
